refactor: update import paths to consistently use data types from the data folder
