+++ src/supabase/SupabaseClientSingleton.ts
@@ -1,6 +1,6 @@
 import { Database } from '@rory0304/molab-project-types';
 import { createClient } from '@supabase/supabase-js';
-import { SupabaseClientType } from 'src/types/supabase';
+import { SupabaseClientType } from 'src/data/types/supabase';
 
 export class SupabaseClientSingleton {
   private static _client?: SupabaseClientType;
