+++ src/utils/editor.ts
@@ -1,5 +1,5 @@
+import { SupabaseClientType } from 'src/data/types/supabase';
 import { ImageUseCase } from 'src/domain/useCases/image';
-import { SupabaseClientType } from 'src/types/supabase';
 
 export const editorImageHandler = ({
   supabase,
